Separated form from popup and also separated form logic/state into a custom React hook
