+++ src/components/Navigation/Navbar.js
@@ -19,8 +19,13 @@ function Navbar() {
           </div>
           <ul className={click ? 'nav-menu active' : 'nav-menu'}>
             <li className="nav-item">
-              <Link to="/test-edit-profile" className="nav-links" onClick={closeMobileMenu}>
+              <Link to="/test-edit-profile-popup" className="nav-links" onClick={closeMobileMenu}>
                 TestEditProfilePopup
+              </Link>
+            </li>
+            <li className="nav-item">
+              <Link to="/test-edit-profile" className="nav-links" onClick={closeMobileMenu}>
+                TestEditProfileForm
               </Link>
             </li>
             <li className="nav-item">
