+++ src/components/Profile/EditProfileForm/index.js
@@ -1,0 +1,3 @@
+import EditProfileForm from './EditProfileForm';
+
+export default EditProfileForm;
